Make it possible to pass an SSLContext to SSL connections
Requires a very unfortunate test faux pas: in order to make the SSL socket connect I need to grab the instance variable of the connection object, I can't come up with another way of simulating a connect. With regular connections it's the socket that is passed to the reactor, but with the SSL socket it's also the socket, and not the SSL socket, but it's the SSL socket you call #connect_nonblock on.