Make a test less specific to support both MRI and JRuby
It's a bit of an ugly test, since it uses the fact that an injected object is passed into OpenSSL, which borks when it gets something it doesn't understand – just to assert that it gets passed in. Turns out MRI and JRuby raise errors with different messages, which shouldn't be very surprising.