Avoid taking locks in resolved and failed Futures
When a Future's state can be seen as not pending we don't need to take the lock to read @value and @error, because we (should) be guaranteed to have seen the writes to those variables.

This makes #value, #on_value and the other read paths approximately 3x faster than when taking a lock.