Expect only two arguments when using #on_complete
The interoperability interface will probably be two arguments, so it’s better to use this internally too. Unfortunately that also means we need to rely on checking for the presence of an error to determine state.

There are some more places where the state is checked explicitly with #resolved? and #failed? so there’s some more work to be done on interoperability.