Rewrite the unblocker to avoid reading and writing
Only a single byte needs to be in it for it to unblock select. This attempts to avoid making additional writes when the unblocker is already unblocked. It also avoids reading when the unblocker has not been unblocked.

I can’t get MRI to work properly, though. For some reason it doesn’t discover that the unblocker is readable with this implementation, so I’ve enabled the write-side of the optimization only for JRuby (where it’s a huge boost).