Don't attempt to lock twice when draining
When the connection is in the "draining" state and the connection isn't
writable it'll attempt to close the connection.
This is done while holding the lock, though the first thing that `#close`
does it to attempt to acquire the lock, which will result in a
ThreadError (Mutex relocking by same thread).
Ultimately we'll end up with a future, and when we call `#value` on said
future we'll get an Ione::Io::ConnectionClosedError exception with
"Mutex relocking by same thread" as the error message.

I'm not at all sure if this is the best solution though.