Add support for global on-listener-error listener
As is, any exception in an either of the future listener callbacks will
just be swallowed. This makes debugging of future-related problems much
harder than it needs to be.

This injects a global listener error handler, which can be helpful when
debugging. Given the motivation, I expect there to be at most one such
listener per application, so this listener behaves a bit differently to
the other future listeners, in that there can only be one listener, and
a second call to the method just replaces the listener with a new one.

This only introduces one level of monitoring, and any exception thrown
by this error handler will still be silently discarded.