Add non-raising variants of Promise#fulfill and Promise#fail
In cases where a promise might be fulfilled or failed concurrently from
multiple threads, it is expected that completing the future might fail
due to already have been completed in another thread.

Even if you query the future for completed status, nothing guarantees
that the future isn't completed between the status check and you next
call. In order to avoid having to rescue the expected exception in this
case, this allows you to just try to fail or fulfill the promise.