Try-complete futures in raise-conditional situations
This changes the internals at various places to use the non-raising
variants in cases where a raise condition is expected. This is mainly
done to make the code easier to understand, as the difference will not
really be externally visible, as most are called from listener blocks
where all errors will be swallowed anyhow. In much the same way, cases
where concurrent calls are not expected are kept the way they are, even
though it (theoretically) incurs a slight overhead to do so.