Close socket for reading when draining
Reading can only happen in the connected stat anyhow, so this just
informs the kernel (and the other end of the connection) of this very
fact.

Not sure if the close_read calls really need to be conditional, as they
seem to be supported both for normal sockets and SSL sockets, but it
doesn't really hurt to have the extra precaution there. They are needed
for the virtual unblocker and acceptor sockets, but that could obviously
be handled differently.