Drain all in-flight socket writes on reactor close
The draining is a simple loop that tries to drain all the sockets until
all write buffers are empty. Technically, this only flushes the
application buffers, and there is no guarantee that the kernel and
network actually manages to get the bytes to the other side.

The draining is done with a timeout (which is currently hard-coded),
and in the event that draining fails, the reactor stop will also fail,
but all the existing shutdown logic will still be performed.

In order to not have to make the drain call conditional, the unblocker
and the acceptors implement drain methods. For the acceptor, the
distinction between drain and close doesn't make any sense, so in this
case it is just a simple alias. While the unblocker shouldn't be needed
during the drain, I wasn't 100% sure that is would be safe, so there
drain does nothing.