Fix a test that deadlocks in JRuby
I think the problem is that the reactor most of the time is not in select when #stop is called, so it just skips to the stopping phase directly. With this test implementation the reactor will be blocked when #stop is called, and the test will kind of test that it continues running until the future resolves.

It's not a great test to say the least.