Reduce lock duration for BaseConnection#flush
This avoids performing the nonblocking write while holding the lock.
While it seems counter-intuitive, it seems this operation frequently
takes a lot of time. I don't know whether it is due to a profiling
artifact or not, but it shouldn't hurt in any case.

Since reads from the write buffer only happen from the single reactor
thread, we could actually perform (some) flushes without locking before
performing the non-blocking write. This uses a new byte-buffer accessor
that allows peeking at the contents of the byte buffer without taking
any lock.