Silence the "using default DH parameters" warning in the HTTP client tests
I'm not sure what the warning means, but returning an OpenSSL::PKey::DH from `:SSLTmpDhCallback` silences it.

The key was generated with `openssl dhparam -out examples/http_client/spec/resources/dh.pem 2048`.

I've also included the error message when the HTTP server fails to start to make it easier to debug these things in the future.